Fix transcription segfault caused by latest updates to pytorch libraries
The stable-ts pip package does not impose a version constraint on its torch-related dependencies.
The newest versions of torch and torchaudio (released a couple of days ago) are not compatible with the ffmpeg version and in general our setup. So by adding this new fixed (!) version after (!) manim-voiceover, we fix the transcription error that occurs when trying to use whisper to get the timestamps for the audio.